fixed the testing bugs and added test consts

+++ src/main.cpp
@@ -7,8 +7,8 @@ using namespace std;
 // Driver code
 int main()
 {
-    std::vector<float> X {1714, 1664, 1760, 1685, 1693, 1670};
-    std::vector<float> Y {2.4, 2.52, 2.54, 2.74, 2.83, 2.91};
+    std::vector<double> X {1714, 1664, 1760, 1685, 1693, 1670};
+    std::vector<double> Y {2.4, 2.52, 2.54, 2.74, 2.83, 2.91};
 
     Regression reg(X, Y);
     reg.fit();
